refactor: rename wav.rs to audio.rs
